+++ components/types.ts
@@ -1,0 +1,6 @@
+export interface Coords {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
